refactor: Moves ALL_PROJECTS and ALL_ENVIRONMENTS to constants
